c 1.0

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,9 +10,9 @@
  * Return: Nothing.
  */
 
-void array_iterator(int *array, int size, void (*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	unsigned int i;
 
 	if (action == NULL)
 	{
